--- src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/BlackCommentResponseDto.kt
@@ -1,5 +1,0 @@
-package com.kotlin.boardproject.domain.comment.dto
-
-class BlackCommentResponseDto(
-    val id: Long,
-)
