+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/black/BlackCommentRequestDto.kt
@@ -1,4 +1,4 @@
-package com.kotlin.boardproject.domain.comment.dto
+package com.kotlin.boardproject.domain.comment.dto.black
 
 import com.kotlin.boardproject.global.enums.BlackReason
 import javax.validation.constraints.NotNull
