+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/black/BlackCommentResponseDto.kt
@@ -1,0 +1,5 @@
+package com.kotlin.boardproject.domain.comment.dto.black
+
+class BlackCommentResponseDto(
+    val id: Long,
+)
