+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/create/CreateCommentRequestDto.kt
@@ -1,13 +1,13 @@
-package com.kotlin.boardproject.domain.comment.dto
+package com.kotlin.boardproject.domain.comment.dto.create
 
 import javax.validation.constraints.NotBlank
 import javax.validation.constraints.NotNull
 import javax.validation.constraints.Positive
 
-data class UpdateCommentRequestDto(
+data class CreateCommentRequestDto(
     @field:NotBlank
     val content: String,
-    @field: NotNull
+    @field:NotNull
     val isAnon: Boolean,
     @field:Positive
     val postId: Long,
