+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/create/CreateCommentResponseDto.kt
@@ -1,4 +1,4 @@
-package com.kotlin.boardproject.domain.comment.dto
+package com.kotlin.boardproject.domain.comment.dto.create
 
 data class CreateCommentResponseDto(
     val id: Long,
