+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/delete/DeleteCommentRequestDto.kt
@@ -1,4 +1,4 @@
-package com.kotlin.boardproject.domain.comment.dto
+package com.kotlin.boardproject.domain.comment.dto.delete
 
 data class DeleteCommentRequestDto(
     val postId: Long,
