+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/delete/DeleteCommentResponseDto.kt
@@ -1,4 +1,4 @@
-package com.kotlin.boardproject.domain.comment.dto
+package com.kotlin.boardproject.domain.comment.dto.delete
 
 data class DeleteCommentResponseDto(
     val id: Long,
