+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/like/LikeCommentResponseDto.kt
@@ -1,4 +1,4 @@
-package com.kotlin.boardproject.domain.comment.dto
+package com.kotlin.boardproject.domain.comment.dto.like
 
 data class LikeCommentResponseDto(
     val id: Long,
