+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/update/UpdateCommentRequestDto.kt
@@ -1,13 +1,13 @@
-package com.kotlin.boardproject.domain.comment.dto
+package com.kotlin.boardproject.domain.comment.dto.update
 
 import javax.validation.constraints.NotBlank
 import javax.validation.constraints.NotNull
 import javax.validation.constraints.Positive
 
-data class CreateCommentRequestDto(
+data class UpdateCommentRequestDto(
     @field:NotBlank
     val content: String,
-    @field:NotNull
+    @field: NotNull
     val isAnon: Boolean,
     @field:Positive
     val postId: Long,
