+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/dto/update/UpdateCommentResponseDto.kt
@@ -1,4 +1,4 @@
-package com.kotlin.boardproject.domain.comment.dto
+package com.kotlin.boardproject.domain.comment.dto.update
 
 data class UpdateCommentResponseDto(
     val id: Long,
