+++ src/main/kotlin/com/kotlin/boardproject/domain/comment/service/CommentService.kt
@@ -1,16 +1,16 @@
 package com.kotlin.boardproject.domain.comment.service
 
-import com.kotlin.boardproject.domain.comment.dto.BlackCommentRequestDto
-import com.kotlin.boardproject.domain.comment.dto.BlackCommentResponseDto
-import com.kotlin.boardproject.domain.comment.dto.CancelLikeCommentResponseDto
-import com.kotlin.boardproject.domain.comment.dto.CreateCommentRequestDto
-import com.kotlin.boardproject.domain.comment.dto.CreateCommentResponseDto
-import com.kotlin.boardproject.domain.comment.dto.DeleteCommentRequestDto
-import com.kotlin.boardproject.domain.comment.dto.DeleteCommentResponseDto
-import com.kotlin.boardproject.domain.comment.dto.LikeCommentResponseDto
-import com.kotlin.boardproject.domain.comment.dto.MyCommentResponseDto
-import com.kotlin.boardproject.domain.comment.dto.UpdateCommentRequestDto
-import com.kotlin.boardproject.domain.comment.dto.UpdateCommentResponseDto
+import com.kotlin.boardproject.domain.comment.dto.black.BlackCommentRequestDto
+import com.kotlin.boardproject.domain.comment.dto.black.BlackCommentResponseDto
+import com.kotlin.boardproject.domain.comment.dto.like.CancelLikeCommentResponseDto
+import com.kotlin.boardproject.domain.comment.dto.create.CreateCommentRequestDto
+import com.kotlin.boardproject.domain.comment.dto.create.CreateCommentResponseDto
+import com.kotlin.boardproject.domain.comment.dto.delete.DeleteCommentRequestDto
+import com.kotlin.boardproject.domain.comment.dto.delete.DeleteCommentResponseDto
+import com.kotlin.boardproject.domain.comment.dto.like.LikeCommentResponseDto
+import com.kotlin.boardproject.domain.comment.dto.read.MyCommentResponseDto
+import com.kotlin.boardproject.domain.comment.dto.update.UpdateCommentRequestDto
+import com.kotlin.boardproject.domain.comment.dto.update.UpdateCommentResponseDto
 import com.kotlin.boardproject.domain.notification.dto.NotificationCreateDto
 import org.springframework.data.domain.Pageable
